+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Added
 - allow disabling Android Lint when no Android plugin is used.
+- detect when the project is a child of a composite build and has no startup tasks.
 
 
 ## [0.10.0] - 2024-05-26